% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_samovar.R
\name{annotation2samovar}
\alias{annotation2samovar}
\title{Process annotation data.frame to SamovaR}
\usage{
annotation2samovar(data)
}
\arguments{
\item{data}{Processed abundance table.
Row names: sequence IDs,
Column names:
\itemize{
\item annotators: (starting with \code{taxID_});
\item \code{true}: for true annotation
\item \code{length}: length of sequence
\item sample
}}
}
\value{
list of samovar data objects
}
\description{
Process annotation data.frame to SamovaR
}
\examples{
library(samovaR)

data <- read_annotation_dir(system.file("extdata", "test_annotations", package = "samovaR"))
gglist <- viz_annotation(data)

samovar_list <- annotation2samovar(data)
}
