% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMrepo_type2data.R
\name{GMrepo_type2data}
\alias{GMrepo_type2data}
\title{Get data from GMrepo}
\usage{
GMrepo_type2data(
  mesh_ids = c("D006262"),
  number_to_process = F,
  number_to_out = F,
  at_level = "species",
  QC_filter = "QCStatus"
)
}
\arguments{
\item{mesh_ids}{Character. All types of meshID to use. List of relations between meshID and phenotype could be obtained using GMrepo_meshID(). Health meshID by default}

\item{number_to_process}{False by default, or maximum number of runs per meshID}

\item{number_to_out}{False by default, maximum number of obtained data}

\item{at_level}{"species" by default. level to obtain classification from GMrepo}

\item{QC_filter}{QCStatus by default. Perform auto QC filtering based on metadata column, or False for no checking.}
}
\description{
Wrapper around GMrepo_type2run and GMrepo_run2data functions
}
\examples{
library(samovaR)

# get data from GMrepo
data_GMrepo <- GMrepo_type2data(mesh_ids = "D006262", number_to_process = 1000)

# equal to:
run_GMrepo <- GMrepo_type2run(mesh_ids = "D006262", number_to_process = 1000)
data_GMrepo <- GMrepo_run2data(run_GMrepo)

# filter runs before obtaining data (OOP updating data!)
run_GMrepo <- filter_samovar(run_GMrepo, "checking", 1)

# view
data_GMrepo

# access to metadata
data_GMrepo$run

# access to data
data_GMrepo$data

# access to runs
data_GMrepo$run

# access to taxa
data_GMrepo$species
}
