% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMrepo_type2run.R
\name{GMrepo_type2run}
\alias{GMrepo_type2run}
\title{Get runs from GMrepo by meshID}
\usage{
GMrepo_type2run(mesh_ids = c("D006262"), number_to_process = F)
}
\arguments{
\item{mesh_ids}{Character. All types of meshID to use. List of relations between meshID and phenotype could be obtained using \code{GMrepo_meshID()}}

\item{number_to_process}{False by default, or maximum number of runs per meshID}
}
\description{
Get runs from GMrepo by meshID
}
\examples{
library(samovaR)

# get data from GMrepo
data_GMrepo <- GMrepo_type2data(mesh_ids = "D006262", number_to_process = 1000)

# equal to:
run_GMrepo <- GMrepo_type2run(mesh_ids = "D006262", number_to_process = 1000)
data_GMrepo <- GMrepo_run2data(run_GMrepo)

# filter runs before obtaining data (OOP updating data!)
run_GMrepo <- filter_samovar(run_GMrepo, "checking", 1)

# view
data_GMrepo

# access to metadata
data_GMrepo$run

# access to data
data_GMrepo$data

# access to runs
data_GMrepo$run

# access to taxa
data_GMrepo$species
}
