% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{phyloseq2samovar}
\alias{phyloseq2samovar}
\title{Convert phyloseq object to samovar object}
\usage{
phyloseq2samovar(phyloseq_data)
}
\arguments{
\item{phyloseq_data}{A phyloseq object}
}
\value{
A samovar object
}
\description{
Convert phyloseq object to samovar object
}
