% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\description{
Re-exported from magrittr so \code{library(samovaR)} is enough for generation
pipelines in vignettes and examples.
}
\keyword{internal}
