% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.samovar.R
\name{read.abundance}
\alias{read.abundance}
\title{Build samovar data object from file or environment}
\usage{
read.abundance(data, metadata)
}
\arguments{
\item{data}{Data.frame or path to abundance file. Row names is using as species list, column names as sample list. Unique names required}

\item{metadata}{Data.frame or path to metadata file}

\item{...}{Parameters processed by read.table()}
}
\description{
Build samovar data object from file or environment
}
