% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_samovar.R
\name{read_annotation_dir}
\alias{read_annotation_dir}
\title{Read annotations produced with samovar pipeline from the directory}
\usage{
read_annotation_dir(data_dir, sample_name_position = 0, ...)
}
\arguments{
\item{data_dir}{Path to abundance table. Row names: sequence IDs, column names: annotators; true for true annotation}

\item{sample_name_position}{Position to split the file basename by "." to extract sample names. 0 by default}

\item{...}{Parameters processed by read.csv()}
}
\value{
samovar object
}
\description{
Read annotations produced with samovar pipeline from the directory
}
\examples{
library(samovaR)

data <- read_annotation_dir(system.file("extdata", "test_annotations", package = "samovaR"))
gglist <- viz_annotation(data)

samovar_list <- annotation2samovar(data)
}
