% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_samovar.R
\name{read_annotation_table}
\alias{read_annotation_table}
\title{Build samovar data object from file or environment}
\usage{
read_annotation_table(data, metadata = F, ...)
}
\arguments{
\item{data}{Data.frame or path to abundance file. Row names is using as species list, column names as sample list. Unique names required}

\item{metadata}{Data.frame or path to metadata file}

\item{...}{Parameters processed by read.csv()}
}
\description{
Build samovar data object from file or environment
}
