% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_samovar.R
\name{read_samovar}
\alias{read_samovar}
\title{Build samovar data object from file or environment}
\usage{
read_samovar(data, metadata = F, ...)
}
\arguments{
\item{data}{path to abundance matrix
Row names: organisms/OTUs/ASVs
Column names: samples}

\item{metadata}{Data.frame or path to metadata file}

\item{...}{Parameters processed by read.csv()}
}
\description{
Build samovar data object from file or environment
}
