% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{samovar2phyloseq}
\alias{samovar2phyloseq}
\title{Convert samovar object to phyloseq object}
\usage{
samovar2phyloseq(samovar_data)
}
\arguments{
\item{samovar_data}{A samovar object}
}
\value{
A phyloseq object
}
\description{
Convert samovar object to phyloseq object
}
