% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{samovar_run}
\alias{samovar_run}
\alias{samovar_run-class}
\alias{$,samovar_run-method}
\alias{$<-,samovar_run-method}
\alias{show,samovar_run-method}
\title{Samovar run data class}
\usage{
\S4method{$}{samovar_run}(x, name)

\S4method{$}{samovar_run}(x, name) <- value

\S4method{show}{samovar_run}(object)
}
\arguments{
\item{x}{A \code{samovar_run} object}

\item{name}{Slot name}

\item{value}{Replacement value}

\item{object}{A \code{samovar_run} object}
}
\description{
Samovar run data class
}
\section{Slots}{

\describe{
\item{\code{metadata}}{metadata DataFrame}

\item{\code{data}}{data}

\item{\code{run}}{character, sample IDs}

\item{\code{cluster}}{character}

\item{\code{species}}{character}
}}

