% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_samovar.R
\name{table2samovar}
\alias{table2samovar}
\title{Build samovar object from the abundance matrix}
\usage{
table2samovar(data, metadata = F, ...)
}
\arguments{
\item{data}{abundance matrix
Row names: organisms/OTUs/ASVs
Column names: samples}

\item{metadata}{metadata data.frame in format: ADD, or FALSE}

\item{...}{data_samovar$rebuild options: min_sp, min_samp}
}
\description{
Build samovar object from the abundance matrix
}
