% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teabag_brew.R
\name{teabag_brew}
\alias{teabag_brew}
\title{Build samovar object}
\usage{
teabag_brew(
  samovar_data,
  dist_function = function(x) dist(x),
  network = F,
  min_cluster_size = 10,
  max_cluster_size = 100,
  plot_log = F,
  ...
)
}
\arguments{
\item{samovar_data}{samovar data after preprocessing stages}

\item{dist_function}{function used for measuring distances between species based on samples}

\item{network}{FALSE or graph that can be used for generation. To be implemented}

\item{min_cluster_size}{FALSE or minimum number of species per cluster}

\item{max_cluster_size}{FALSE or maximum number of species per cluster}

\item{plot_log}{Logical or path for log plots output}

\item{...}{Additional arguments, passed}
}
\description{
Samovar network is a 2D-oriented graph with metadata and abundances of species per sample. Oriented graph could be used for network prediction, or for better generation some network could be used as initial (to be implemented)
For better understanding of building database and using it in generation, visit github source
}
