% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tealeaves_pack.R
\name{tealeaves_pack}
\alias{tealeaves_pack}
\title{Scale species abundances}
\usage{
tealeaves_pack(
  samovar_data,
  normalization_function = function(x) log10(x + 1),
  plot_log = T,
  ...
)
}
\arguments{
\item{samovar_data}{Samovar data object to rescale}

\item{normalization_function}{Function using for rescaling}

\item{plot_log}{Logical or path for log plots output}

\item{...}{Additional arguments, passed}
}
\description{
Scale species abundances
}
\examples{
library(samovaR)

# download data
teatree <- GMrepo_type2data(number_to_process = 2000)

# filter
tealeaves <- teatree \%>\%
  teatree_trim(treshhold_species = 3, treshhold_samples = 3, treshhold_amount = 10^(-3))

# normalizing
## if you build teatree by your own, rescaling stage when building via teatree$rescale()
## or assigning teatree$min_value and teatree$max_value is required
## good approximation to normal distribution is required for glm generating methods
teabag <- tealeaves \%>\%
  tealeaves_pack(normalization_function = function(x) log10(x+1))

# clustering
concotion <- teabag \%>\%
  teabag_brew(min_cluster_size = 4, max_cluster_size = 6)
# remember: if you want to refilter, it is better to re-do welding stage to avoid crashes in future!

# building samovar
samovar <- concotion \%>\%
  concotion_pour()
}
