% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teatree_trim.R
\name{teatree_trim}
\alias{teatree_trim}
\title{Filter species and samples from samovar_data object}
\usage{
teatree_trim(
  samovar_data,
  metadata_filter = F,
  treshhold_amount = 10^(-5),
  treshhold_samples = 1,
  treshhold_species = 1,
  drop_species = F,
  drop_unclassified = T,
  ...
)
}
\arguments{
\item{samovar_data}{Samovar data object to filter}

\item{metadata_filter}{False, character or data.frame with 2 columns: first contain metadata names for filtering, and second values per column}

\item{treshhold_amount}{Minimum value to conclude as not the noise.}

\item{treshhold_samples}{Minimum number of representing samples to keep species.}

\item{treshhold_species}{Minimum number of representing species to keep samples.}

\item{drop_species}{Species names to drop, or FALSE}

\item{drop_unclassified}{Drop unknown and unclassified ranks. True by default}

\item{...}{Additional arguments, passed}
}
\description{
Filter species and samples from samovar_data object
}
\examples{
library(samovaR)

# download data
teatree <- GMrepo_type2data(number_to_process = 2000)

# filter
tealeaves <- teatree \%>\%
  teatree_trim(treshhold_species = 3, treshhold_samples = 3, treshhold_amount = 10^(-3))

# normalizing
## if you build teatree by your own, rescaling stage when building via teatree$rescale()
## or assigning teatree$min_value and teatree$max_value is required
## good approximation to normal distribution is required for glm generating methods
teabag <- tealeaves \%>\%
  tealeaves_pack(normalization_function = function(x) log10(x+1))

# clustering
concotion <- teabag \%>\%
  teabag_brew(min_cluster_size = 4, max_cluster_size = 6)
# remember: if you want to refilter, it is better to re-do welding stage to avoid crashes in future!

# building samovar
samovar <- concotion \%>\%
  concotion_pour()
}
