% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_config.R
\name{unpack_config}
\alias{unpack_config}
\title{Unpack SAMOVAR config}
\usage{
unpack_config(config_samovar)
}
\arguments{
\item{config_samovar}{SAMOVAR config}
}
\value{
arguments for samovar_preprocess and samovar_boil
}
\description{
Unpack SAMOVAR config to a list of parameters
}
\examples{
library(samovaR)

# Example config
tf <- tempfile()

yaml::write_yaml(
  list(
    treshhold_amount = 10^(-5),
    plot_log = F,
    min_cluster_size = 5,
    N = 5,
    N_reads = 100
  ),
  tf
)

config_samovar <- unpack_config(tf)

## build samovar directly from the table
data <- matrix(
  rlnorm(625),
  dimnames = list(
    rownames = paste0("sp_", 1:25),
    colnames = paste0("sample_", 1:25)
  ),
  nrow = 25)

samovar_data <- table2samovar(data)

# Run with config
config_samovar$samovar_data <- samovar_data
samovar <- do.call(samovar_preprocess, config_samovar)
samovar_new <- samovar_boil(samovar, N = config_samovar$N)

new_data <- samovar_new$data * config_samovar$N_reads

heatmap(as.matrix(new_data))
}
