% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_composition.R
\name{viz_composition}
\alias{viz_composition}
\title{Visualize composition}
\usage{
viz_composition(
  data,
  reord_samples = "fpc",
  reord_species = "amount",
  type = "column",
  top = 15,
  interactive = F,
  ggplot_add = F,
  bottom_legend = F
)
}
\arguments{
\item{data}{data.frame with dimensions of species * samples, or samovar objects: samovar_data, samovar_base, samovar_run or GMrepo_run (with data)}

\item{reord_samples}{character, fpc, fpc_scaled, hcl, amount, tsne, or none reorder of samples on plot}

\item{reord_species}{character, same for reord_samples}

\item{type}{character, column or tile for composition visualize, or donut (to implement) for mean composition visualization}

\item{top}{integer, number of top-represented taxa to show, or FALSE to show all}

\item{interactive}{logical. ggplot or plotly object to return}

\item{ggplot_add}{functions to add to ggplot object, or FALSE.}

\item{bottom_legend}{vector length of samples to show on plot as a color legend, or FALSE}
}
\description{
Visualize composition
}
\examples{
library(samovaR)

# Download data
teatree <- GMrepo_type2data(number_to_process = 1000)

#Composition
viz_composition(teatree)
}
