% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_samovar.R
\name{annotation2samovar}
\alias{annotation2samovar}
\title{Process annotation data.frame to SamovaR}
\usage{
annotation2samovar(data)
}
\arguments{
\item{data}{Processed abundance table. Row names: sequence IDs, column names: annotators; true for true annotation}
}
\value{
list of samovar data objects
}
\description{
Process annotation data.frame to SamovaR
}
\examples{
library(samovaR)

data <- read_annotation_dir("data/test_annotations/")
gglist <- viz_annotation(data)
}
