% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_samova.R, R/build_samovar_deprecated.R
\name{build_samovar}
\alias{build_samovar}
\title{Build samovar object}
\usage{
build_samovar(
  samovar_data,
  dist_function = function(x) dist(x),
  network = F,
  k_means = F,
  min_cluster_size = F,
  plot_log = T
)

build_samovar(
  samovar_data,
  dist_function = function(x) dist(x),
  network = F,
  k_means = F,
  min_cluster_size = F,
  plot_log = T
)
}
\arguments{
\item{samovar_data}{samovar data after preprocessing stages}

\item{network}{FALSE or graph that can be used for generation. To be implemented}

\item{plot_log}{Logical or path for log plots output}

\item{distance_function}{function used for measuring distances between species based on samples}

\item{min_min_cluster_size}{FALSE or minimum number of species per cluster}

\item{max_min_cluster_size}{FALSE or minimum number of species per cluster}
}
\description{
Samovar network is a 2D-oriented graph with metadata and abundances of species per sample. Oriented graph could be used for network prediction, or for better generation some network could be used as initial (to be implemented)
For better understanding of building database and using it in generation, visit github source

Samovar network is a 2D-oriented graph with metadata and abundances of species per sample. Oriented graph could be used for network prediction, or for better generation some network could be used as initial (to be implemented)
For better understanding of building database and using it in generation, visit github source
}
