% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{GMrepo_run}
\alias{GMrepo_run}
\title{GMrepo run data class}
\description{
GMrepo run data class
}
\section{Slots}{

\describe{
\item{\code{description}}{metadata DataFrame}

\item{\code{run}}{character}
}}


