% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{GMrepo_type2data_or_example}
\alias{GMrepo_type2data_or_example}
\title{Download GMrepo data with an offline fallback}
\usage{
GMrepo_type2data_or_example(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{GMrepo_type2data}}.}
}
\value{
A \code{samovar_data} object.
}
\description{
Download GMrepo data with an offline fallback
}
