% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_samovar.R
\name{read_annotation_dir}
\alias{read_annotation_dir}
\title{Build samovar annotation object from directory}
\usage{
read_annotation_dir(data_dir, metadata = F, ...)
}
\arguments{
\item{data_dir}{Path to abundance table. Row names: sequence IDs, column names: annotators; true_annotation}

\item{metadata}{Data.frame or path to metadata file}

\item{...}{Parameters processed by read.csv()}
}
\description{
Build samovar annotation object from directory
}
