% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_samovar.R
\name{read_samovar}
\alias{read_samovar}
\title{Build samovar data object from file or environment}
\usage{
read_samovar(data, metadata = F, ...)
}
\arguments{
\item{data}{path to abundance matrix
Row names: organisms/OTUs/ASVs
Column names: samples}

\item{metadata}{Data.frame or path to metadata file}

\item{...}{Parameters processed by read.csv()}
}
\value{
samovar object
}
\description{
Build samovar data object from file or environment
}
\examples{
library(samovaR)

## build samovar directly from the table
data <- matrix(
  runif(25),
  dimnames = list(
    rownames = paste0("sp_", 1:5),
    colnames = paste0("sample_", 1:5)
  ),
  nrow = 5)

samovar_data <- table2samovar(data)
print(samovar_data)

# or read abundance table
tf <- tempfile()
write.csv(data, tf)

samovar_data <- read_samovar(tf)
print(samovar_data)
}
