% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samovaR-package.R
\docType{package}
\name{samovaR-package}
\alias{samovaR}
\alias{samovaR-package}
\title{samovaR: R package for generating model metagenomes with specified properties}
\description{
There is a fundamental problem in modern metagenomics: there are huge differences between methodological approaches that strongly influence the results, while remaining outside the attention of researchers. We propose an approach that utilizes de novo generation of the artificial metagenomes - SamovaR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dsmutin/samovar/}
}

}
\keyword{internal}
