% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samovaR-package.R
\docType{package}
\name{samovaR-package}
\alias{samovaR}
\alias{samovaR-package}
\title{samovaR: R package for generating model metagenomes with specified properties}
\description{
Metagenomics serves as a fundamental approach in the analysis of biological communities. The field continuously witnesses the emergence of numerous novel tools, which in turn necessitates the validation of these tools to address the crucial challenge at hand. In light of this, we have devised an artificial data generation tool `samovaR`, aimed at enhancing the development of algorithms and expediting scientific discoveries. This addon implement creation of abundance files
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dsmutin/samovar/tree/beta}
}

}
\keyword{internal}
