% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{bind_samovar}
\alias{bind_samovar}
\alias{filter_samovar}
\alias{add_species}
\alias{export_data}
\alias{get_cluster_data}
\alias{rescale_samovar}
\alias{normalize_samovar}
\alias{reverse_normalize_samovar}
\alias{reverse_normalize_df_samovar}
\alias{rebuild_samovar}
\alias{get_clean_samovar}
\alias{get_data_samovar}
\alias{cluster_count}
\alias{get_cluster_info}
\alias{get_cluster_length}
\alias{get_max_inter_cluster}
\alias{get_max_inner_cluster}
\alias{bind_samovar,GMrepo_run-method}
\alias{filter_samovar,GMrepo_run-method}
\alias{bind_samovar,samovar_run-method}
\alias{add_species,samovar_run-method}
\alias{filter_samovar,samovar_run-method}
\alias{get_cluster_data,samovar_run-method}
\alias{export_data,samovar_run-method}
\alias{bind_samovar,samovar_data-method}
\alias{filter_samovar,samovar_data-method}
\alias{rescale_samovar,samovar_data-method}
\alias{normalize_samovar,samovar_data-method}
\alias{reverse_normalize_samovar,samovar_data-method}
\alias{reverse_normalize_df_samovar,samovar_data-method}
\alias{rebuild_samovar,samovar_data-method}
\alias{get_clean_samovar,samovar_data-method}
\alias{get_data_samovar,samovar_data-method}
\alias{cluster_count,samovar_data-method}
\alias{get_cluster_info,samovar_base-method}
\alias{get_cluster_length,samovar_base-method}
\alias{get_max_inter_cluster,samovar_base-method}
\alias{get_max_inner_cluster,samovar_base-method}
\title{S4 generics for samovaR objects}
\usage{
bind_samovar(object, bind_with)

filter_samovar(object, field, value)

add_species(object, cluster_sp, sp, abundance, to_run = 1)

export_data(object)

get_cluster_data(object, cl)

rescale_samovar(object, scale_function = function(x) x/sum(x))

normalize_samovar(
  object,
  normalization_function = object@normalization_function
)

reverse_normalize_samovar(object)

reverse_normalize_df_samovar(object, x)

rebuild_samovar(object, min_sp = 1, min_samp = 1)

get_clean_samovar(object, cl)

get_data_samovar(object, cl)

cluster_count(object)

get_cluster_info(object, sp)

get_cluster_length(object, cl)

get_max_inter_cluster(object, predicted, type = "prob")

get_max_inner_cluster(object, cl, predicted, type = "prob")

\S4method{bind_samovar}{GMrepo_run}(object, bind_with)

\S4method{filter_samovar}{GMrepo_run}(object, field, value)

\S4method{bind_samovar}{samovar_run}(object, bind_with)

\S4method{add_species}{samovar_run}(object, cluster_sp, sp, abundance, to_run = 1)

\S4method{filter_samovar}{samovar_run}(object, field, value)

\S4method{get_cluster_data}{samovar_run}(object, cl)

\S4method{export_data}{samovar_run}(object)

\S4method{bind_samovar}{samovar_data}(object, bind_with)

\S4method{filter_samovar}{samovar_data}(object, field, value)

\S4method{rescale_samovar}{samovar_data}(
  object,
  scale_function = function(x) {
     x/sum(x)
 }
)

\S4method{normalize_samovar}{samovar_data}(
  object,
  normalization_function = object@normalization_function
)

\S4method{reverse_normalize_samovar}{samovar_data}(object)

\S4method{reverse_normalize_df_samovar}{samovar_data}(object, x)

\S4method{rebuild_samovar}{samovar_data}(object, min_sp = 1, min_samp = 1)

\S4method{get_clean_samovar}{samovar_data}(object, cl)

\S4method{get_data_samovar}{samovar_data}(object, cl)

\S4method{cluster_count}{samovar_data}(object)

\S4method{get_cluster_info}{samovar_base}(object, sp)

\S4method{get_cluster_length}{samovar_base}(object, cl)

\S4method{get_max_inter_cluster}{samovar_base}(object, predicted, type = "prob")

\S4method{get_max_inner_cluster}{samovar_base}(object, cl, predicted, type = "prob")
}
\arguments{
\item{object}{A samovaR S4 object}

\item{bind_with}{Object to bind to \code{object}}

\item{field}{Metadata column name}

\item{value}{Filter value}

\item{cluster_sp}{Cluster label for new species}

\item{sp}{Species names}

\item{abundance}{Numeric abundances}

\item{to_run}{Generated-sample column index}

\item{cl}{Cluster identifier}

\item{scale_function}{Column rescaling function}

\item{normalization_function}{Abundance transform}

\item{x}{Matrix or data frame to reverse-normalize}

\item{min_sp}{Minimum species per sample}

\item{min_samp}{Minimum samples per species}

\item{predicted}{Indices already generated}

\item{type}{Graph type, usually \code{"prob"}}
}
\description{
S4 generics for samovaR objects
}
