% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{samovar_base}
\alias{samovar_base}
\alias{samovar_base-class}
\alias{$,samovar_base-method}
\alias{$<-,samovar_base-method}
\alias{show,samovar_base-method}
\title{samovar base class}
\usage{
\S4method{$}{samovar_base}(x, name)

\S4method{$}{samovar_base}(x, name) <- value

\S4method{show}{samovar_base}(object)
}
\arguments{
\item{x}{A \code{samovar_base} object}

\item{name}{Slot name}

\item{value}{Replacement value}

\item{object}{A \code{samovar_base} object}
}
\description{
samovar base class
}
\section{Slots}{

\describe{
\item{\code{samovar_data}}{samovar_data object}

\item{\code{inner_cluster_graph_method}}{list of graphs in matrix form of inner cluster connections}

\item{\code{inter_cluster_graph_method}}{list of graphs in matrix form of inter cluster connections}

\item{\code{inner_cluster_graph_prob}}{list of co-occurrence probabilities in matrix form of inner cluster members}

\item{\code{inter_cluster_graph_prob}}{list of co-occurrence probabilities in matrix form between clusters}

\item{\code{preferences}}{concotion_pour() properties}
}}

