% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samovar_boil.R
\name{samovar_boil}
\alias{samovar_boil}
\title{Generate artificial data}
\usage{
samovar_boil(
  samovar_base,
  N = 1,
  init_sp = F,
  init_ab = F,
  avoid_zero_generations = T,
  seed = 42
)
}
\arguments{
\item{samovar_base}{samovar data after preprocessing and building stages}

\item{N}{number of artificial samples to generate}

\item{init_sp}{species vector for initializing data generation, or FALSE for usage most common taxa, auto for choosing random taxa}

\item{init_ab}{species amount vector (values from 0 to 1) for initializing data generation, or FALSE for mean initial taxa assignment, or auto for usage from known edf for each species from init_sp}

\item{avoid_zero_generations}{logical, avoid zero-based generations or not. FALSE might results in under-distributed communities, while TRUE in over-represented with species from different clusters possibly come from different samples groups}

\item{seed}{initial seed for the seeds generation}
}
\description{
Use pre-built samovar_data with its parameters
}
\examples{
library(samovaR)
library(tidyverse)

# download and prepare data
samovar_raw <- GMrepo_type2data(number_to_process = 2000)

samovar_data <- samovar_raw \%>\%
  samovar_preprocess()

# Similar to:
## samovar_raw \%>\%
##   teatree_trim() \%>\%
##   tealeaves_pack() \%>\%
##   teabag_brew() \%>\%
##   concotion_pour()

# generate
new_data <- samovar_data \%>\%
  samovar_boil(N = 100)
}
