% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{samovar_data}
\alias{samovar_data}
\alias{samovar_data-class}
\title{samovar data class}
\description{
samovar data class
}
\section{Slots}{

\describe{
\item{\code{description}}{metadata DataFrame}

\item{\code{data}}{DataFrame with species abundances. No NA pass}

\item{\code{run}}{character, runs}

\item{\code{species}}{character, runs}

\item{\code{normalization_function}}{normalization function for samples}

\item{\code{reverse_normalization_function}}{reverse normalization function}

\item{\code{min_value}}{minimal value after scaling}

\item{\code{max_value}}{maximal value after scaling}

\item{\code{cluster}}{character vector, enumerated clusters for each species}

\item{\code{cluster_size}}{named numeric, cluster sizes per cluster}
}}

