% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{samovar_data}
\alias{samovar_data}
\title{samovar data class}
\description{
samovar data class

samovar base class
}
\section{Slots}{

\describe{
\item{\code{description}}{metadata DataFrame}

\item{\code{data}}{DataFrame with species abundances. No NA pass}

\item{\code{run}}{character, runs}

\item{\code{species}}{character, runs}

\item{\code{samovar_dataa}}{samovar_data object}

\item{\code{method}}{method to obtain samovar_data}

\item{\code{inner_cluster_graph_method}}{list of graphs in matrix form of inner cluster connections}

\item{\code{inter_cluster_graph_method}}{list of graphs in matrix form of inter cluster connections}

\item{\code{inner_cluster_graph_prob}}{list of co-occurrence probabilities in matrix form of inner cluster members}

\item{\code{inter_cluster_graph_prob}}{list of co-occurrence probabilities in matrix form between clusters}
}}


