% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samovar_preprocess.R
\name{samovar_preprocess}
\alias{samovar_preprocess}
\title{Preprocess SAMOVAR data}
\usage{
samovar_preprocess(
  samovar_data,
  metadata_filter = F,
  treshhold_amount = 10^(-5),
  treshhold_samples = 1,
  treshhold_species = 1,
  drop_species = F,
  drop_unclassified = T,
  normalization_function = function(x) log10(x + 1),
  plot_log = T,
  dist_function = function(x) dist(x),
  network = F,
  min_cluster_size = 2,
  max_cluster_size = 100,
  inner_method = "glm",
  inter_method = "glm",
  inner_model = "gaussian",
  inter_model = "gaussian",
  probability_calculation = "oriented",
  cluster_connection = "mean",
  ...
)
}
\arguments{
\item{samovar_data}{samovar data object}

\item{metadata_filter}{False, character or data.frame with 2 columns: first contain metadata names for filtering, and second values per column}

\item{treshhold_amount}{Minimum value to conclude as not the noise.}

\item{treshhold_samples}{Minimum number of representing samples to keep species.}

\item{treshhold_species}{Minimum number of representing species to keep samples.}

\item{drop_unclassified}{Drop unknown and unclassified ranks. True by default}

\item{normalization_function}{Function using for rescaling}

\item{plot_log}{Logical or path for log plots output}

\item{network}{FALSE or graph that can be used for generation. To be implemented}

\item{min_cluster_size}{FALSE or minimum number of species per cluster}

\item{max_cluster_size}{FALSE or minimum number of species per cluster}

\item{inner_method}{Character, glm, other to be implemented (bootstrap, bsPCA)}

\item{inter_method}{Character, glm, other to be implemented (bootstrap, bsPCA)}

\item{inner_model}{Character, model processed by glm(). For glm mode only. quasipoisson by default}

\item{inter_model}{Character, model processed by glm(). For glm mode only. quasipoisson by default}

\item{cluster_connection}{Character (mean, median), or function. The way of cluster connection. If function, way of summarize all samples of species cluster}

\item{...}{Additional arguments, passed}
}
\value{
Build SAMOVAR object
}
\description{
Wrapper for SAMOVAR preprocess commands
}
\examples{
library(samovaR)
library(tidyverse)

# download and prepare data
samovar_raw <- GMrepo_type2data(number_to_process = 2000)

samovar_data <- samovar_raw \%>\%
  samovar_preprocess()

# Similar to:
## samovar_raw \%>\%
##   teatree_trim() \%>\%
##   tealeaves_pack() \%>\%
##   teabag_brew() \%>\%
##   concotion_pour()

# generate
new_data <- samovar_data \%>\%
  samovar_boil(N = 100)
}
