% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{samovar_run}
\alias{samovar_run}
\alias{samovar_run-class}
\title{Samovar run data class}
\description{
Samovar run data class
}
\section{Slots}{

\describe{
\item{\code{metadata}}{metadata DataFrame}

\item{\code{data}}{data}

\item{\code{run}}{character, sample IDs}

\item{\code{cluster}}{character}

\item{\code{species}}{character}
}}

