% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{samovar_run}
\alias{samovar_run}
\title{Samovar run data class}
\description{
Samovar run data class
}
\section{Slots}{

\describe{
\item{\code{metadata}}{metadata DataFrame}

\item{\code{data}}{data}

\item{\code{run}}{character, samle IDs}
}}

\section{Methods}{

\describe{
\item{\code{export(Class)}}{Returns the result of coercing the object to
Class.  No effect on the object itself.}
}}

