% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_annotation.R
\name{viz_annotation}
\alias{viz_annotation}
\title{Visualize annotation results}
\usage{
viz_annotation(
  data,
  type = c("f1", "R2", "cv", "conf"),
  show_top = 10,
  output_dir = NULL,
  plot = T,
  split = T
)
}
\arguments{
\item{data}{Processed abundance table.
Row names: sequence IDs,
Column names:
\itemize{
\item annotators: (starting with \code{taxID_});
\item \code{true}: for true annotation
\item \code{length}: length of sequence
\item sample
}}

\item{type}{character vector.
\itemize{
\item if present column true_annotation: could be one of
\itemize{
\item "f1",
\item "R2",
\item "confidence",
\item "cross-validation",
\item or their combination (e.g. c("f1", "R2", "cv", "conf"))
}
}}

\item{show_top}{integer. Number of top annotations to show.}

\item{output_dir}{character. Directory to save the plots. If NULL, plots are not saved.}

\item{plot}{logical. If TRUE, plots are printed.}

\item{split}{logical. If TRUE, plots are split into separate files.}
}
\value{
list of ggplot objects
}
\description{
Visualize annotation results
}
\examples{
library(samovaR)

data <- read_annotation_dir("data/test_annotations/")
gglist <- viz_annotation(data)

samovar_list <- annotation2samovar(data)
}
